refactor: Consolidate AuthRepository and DoorRepository to domain interfaces
Delete duplicate interface definitions from androidApp. Both
AuthRepository and DoorRepository had identical signatures in
androidApp and domain module. Now there is one source of truth:
domain/repository/.

- Delete interface from auth/AuthRepository.kt (keep impl + Hilt module)
- Delete interface from door/DoorRepository.kt (keep impl + Hilt module)
- Update all imports to com.chriscartland.garage.domain.repository.*
- Update fakes and tests to use domain interfaces

This unblocks UseCase extraction — UseCases can now depend on domain
interfaces without pulling in androidApp.
